# docs/swagger.json
{"swagger":"2.0","info":{"description":"Hendrix API is allows CRUD operations for Rules, Tenants and Templates in Hendrix","title":"Hendrix API","license":{"name":"The Apache Software License, Version 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0.txt"}},"basePath":"/api","paths":{"/rules/{tenantId}":{"get":{"summary":"List rules","description":"Will return list of rules for a given Tenant ID","operationId":"listRules","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"pretty","in":"query","required":false,"type":"boolean","default":"false"},{"name":"filter","in":"query","required":false,"type":"integer","default":"0","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/Rule"}}}}},"post":{"summary":"Create rule","description":"Will create an empty rule for a given Tenant ID","operationId":"createRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"integer","format":"int32"}}}},"delete":{"summary":"Delete all rules","description":"Delete all rules for the supplied Tenant ID","operationId":"deleteAllRules","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/rules/{tenantId}/disable":{"put":{"summary":"Disable all rules","description":"Disable all rules for the supplied Tenant ID","operationId":"disableAllRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/rules/{tenantId}/{ruleId}":{"get":{"summary":"Get rule","description":"Get rule for the supplied Tenant ID and Rule ID","operationId":"getRule","consumes":["application/json"],"produces":["text/html"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"ruleId","in":"path","required":true,"type":"integer","format":"int32"},{"name":"pretty","in":"query","required":false,"type":"boolean","default":"false"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Rule"}}}},"put":{"summary":"Update rule","description":"Will update rule logic for a given Tenant ID and Rule ID","operationId":"putRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"ruleId","in":"path","required":true,"type":"integer","format":"int32"},{"name":"Accept-Charset","in":"header","required":false,"type":"string","default":""}],"responses":{"200":{"description":"successful operation","schema":{"type":"integer","format":"int32"}}}},"delete":{"summary":"Delete rule","description":"Delete rule for the supplied Tenant ID and Rule ID","operationId":"deleteRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"ruleId","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"default":{"description":"successful operation"}}}},"/rules/{tenantId}/{ruleId}/disable":{"put":{"summary":"Disable rule","description":"Disable rule for the supplied Tenant ID and Rule ID","operationId":"disableRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"ruleId","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Rule"}}}}},"/rules/{tenantId}/{ruleId}/enable":{"put":{"summary":"Enable rule","description":"Enable rule for the supplied Tenant ID and Rule ID","operationId":"enableRule","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"ruleId","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/Rule"}}}}},"/templates/{tenantId}":{"get":{"summary":"List templates","description":"List templates for the supplied Tenant ID","operationId":"listTemplates","produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"pretty","in":"query","required":false,"type":"boolean","default":"false"}],"responses":{"200":{"description":"successful operation","schema":{"type":"array","items":{"$ref":"#/definitions/AlertTemplate"}}}}},"post":{"summary":"Create template","description":"Create an empty for the supplied Tenant ID","operationId":"createTemplate","produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"}],"responses":{"200":{"description":"successful operation","schema":{"type":"integer","format":"int32"}}}},"delete":{"summary":"Delete all templates","description":"Delete all templates for the supplied Tenant ID, templates can only be deleted if there are no rules associated with them","operationId":"deleteAllTemplates","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"}],"responses":{"default":{"description":"successful operation"}}}},"/templates/{tenantId}/{test}":{"get":{"summary":"Get template","description":"Get template for the supplied Tenant ID and Template ID","operationId":"getTemplate","consumes":["application/json"],"produces":["text/html"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"test","in":"path","required":true,"type":"integer","format":"int32"},{"name":"pretty","in":"query","required":false,"type":"boolean","default":"false"}],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/AlertTemplate"}}}},"put":{"summary":"Update template","description":"Update template for the supplied Tenant ID and Template ID","operationId":"putTemplate","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"test","in":"path","required":true,"type":"integer","format":"int32"},{"name":"Accept-Charset","in":"header","required":false,"type":"string","default":""}],"responses":{"200":{"description":"successful operation","schema":{"type":"integer","format":"int32"}}}},"delete":{"summary":"Delete template","description":"Delete template for the supplied Tenant ID and Template ID, template can only be deleted if there are no rules associated with them","operationId":"deleteTemplate","consumes":["application/json"],"produces":["application/json"],"parameters":[{"name":"tenantId","in":"path","required":true,"type":"string"},{"name":"test","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"default":{"description":"successful operation"}}}}},"definitions":{"Condition":{"type":"object"},"Action":{"properties":{"actionType":{"type":"string","enum":["AGGREGATION","STATE","TAG","NEW","RAW_ALERT","TEMPLATED_ALERT","OMEGA","ANOMD"]},"actionId":{"type":"integer","format":"int32"}}},"AlertTemplate":{"properties":{"templateId":{"type":"integer","format":"int32"},"templateName":{"type":"string"},"destination":{"type":"string"},"media":{"type":"string"},"subject":{"type":"string"},"body":{"type":"string"},"throttleDuration":{"type":"integer","format":"int32"},"throttleLimit":{"type":"integer","format":"int32"}}},"Rule":{"properties":{"ruleId":{"type":"integer","format":"int32"},"name":{"type":"string"},"active":{"type":"boolean","default":false},"description":{"type":"string"},"actions":{"type":"array","items":{"$ref":"#/definitions/Action"}},"condition":{"$ref":"#/definitions/Condition"}}}}}